+++ NEWS.md
@@ -2,10 +2,17 @@
 
 ##### Version 1.0.2 (Release date: XXXX-XX-XX)
 
+## New functions
+
+* sort_colms: sort columns in a more intuitive order
+* match_wav_case: corrects the case of the extension name of sound files
+
 ## Changes and additions:
 
+* 'rm_dup' argument to remove duplicated rows in 'imp_raven' function
+* Small changes in vignette table appereance
 * Parallel and progress bar available in 'imp_raven' function
-
+* 'hz.to.khz' and 'khz.to.hz' arguments available in 'relabel_colms' function
 
 ##### Version 1.0.1 (Release date: 2018-02-11)